requirements.txt
# Core
streamlit>=1.32.0
python-dotenv>=1.0.0
supabase==2.15.3
pandas>=2.2.0

# LLM Providers
groq==0.11.0
google-generativeai==0.8.3
mistralai==1.2.3

# Document parsing
PyMuPDF==1.24.13
python-docx==1.1.2

# Image support
Pillow>=11.0.0

# Charts
plotly==5.24.1

psycopg2-binary>=2.9.9
